+++ templates/boilerplate-template/vite.config.ts
@@ -1,6 +1,7 @@
 import path from "path";
 import react from "@vitejs/plugin-react";
 import { defineConfig } from "vite";
+import { VitePWA } from "vite-plugin-pwa";
 
 export default defineConfig({
   build: {
@@ -10,7 +11,35 @@ export default defineConfig({
     open: true,
   },
   plugins: [
-    react()
+    react(),
+    VitePWA({
+      registerType: "autoUpdate",
+      manifest: {
+        name: "Aptos Fullstack Template",
+        short_name: "Aptos Template",
+        icons: [
+          {
+            src: "/icons/icon-192x192.png",
+            sizes: "192x192",
+            type: "image/png",
+            purpose: "any maskable",
+          },
+          {
+            src: "/icons/icon-384x384.png",
+            sizes: "384x384",
+            type: "image/png",
+          },
+          {
+            src: "/icons/icon-512x512.png",
+            sizes: "512x512",
+            type: "image/png",
+          },
+        ],
+        start_url: "/",
+        display: "standalone",
+        orientation: "portrait",
+      },
+    }),
   ],
   resolve: {
     alias: {