+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the create-aptos-dapp tool will be captured in this file.
 # Unreleased
 
 - [Fix] Generate Boilerplate template without Surf errors out because of a wrong file path
+- Support PWA in the boilerplate template
 
 # 0.0.36 (2024-10-28)